requirements.txt
Flask
Pillow
scipy
numpy
scikit-learn
requests>=2.20.0
